+++ packages/website/src/pages/index/custom_section/Lix.tsx
@@ -1,4 +1,5 @@
 import { Button } from "../components/Button.jsx"
+import * as m from "@inlang/paraglide-js/inlang-marketplace/messages"
 
 const Lix = () => {
 	return (
@@ -11,15 +12,15 @@ const Lix = () => {
 			<img class="lg:hidden" src="./images/lixCoverV02.webp" alt="Lix Change Control" />
 			<div class="flex flex-col md:flex-row items-start md:items-end px-8 pb-6 pt-3">
 				<div class="flex flex-col gap-2 flex-1">
-					<h1 class="text-md text-surface-700 font-semibold">Lix change control</h1>
-					<p class="text-sm text-surface-500">⏱️ The backbone of the ecosystem</p>
+					<h1 class="text-md text-surface-700 font-semibold">{m.home_lix_title()}</h1>
+					<p class="text-sm text-surface-500">{m.home_lix_description()}</p>
 				</div>
 				<Button
 					type="textPrimary"
 					href="https://github.com/inlang/monorepo/blob/ff2429c62ca5c4aa9443f18ca66f63249668784f/lix/design-principles.md#L1"
 					class="-mb-[10px]"
 				>
-					{"More about Lix >"}
+					{m.home_lix_button()}
 				</Button>
 			</div>
 		</div>