+++ packages/plugins/json/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @inlang/plugin-json
 
+## 4.9.0
+
+### Minor Changes
+
+- 2150b4873: fix: path patterns can start as as an absolute path like `/resources/{languageTag}.json`
+
+## 4.8.0
+
+### Minor Changes
+
+- 2f924df32: added Modulesettings validation via the Typebox JSON Schema Validation. This ensure that users can exclusively use module settings when there are given by the moduel
+
+### Patch Changes
+
+- Updated dependencies [2f924df32]
+  - @inlang/sdk@0.16.0
+
 ## 4.7.0
 
 ### Minor Changes