+++ packages/versioned-interfaces/plugin/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inlang/plugin
 
+## 2.2.0
+
+### Minor Changes
+
+- 2f924df32: added Modulesettings validation via the Typebox JSON Schema Validation. This ensure that users can exclusively use module settings when there are given by the moduel
+
 ## 2.1.0
 
 ### Minor Changes